+++ pymc3/__init__.py
@@ -3,7 +3,7 @@ __version__ = "3.0"
 from .blocking import *
 from .distributions import *
 from .external import *
-from .linear_models import *
+from . import *
 from .math import logsumexp, logit, invlogit
 from .model import *
 from .stats import *
