+++ pymc3/glm/__init__.py
