+++ pymc3/glm/families.py
