+++ pymc3/glm/utils.py
